[geombong/issue-tracker#13] Refactor: separate input types into another file
