+++ FE/src/type/input.type.ts
@@ -1,0 +1,10 @@
+export type InputStyleProps = {
+  inputStyle: 'large' | 'medium' | 'small';
+};
+
+export type InputProps = InputStyleProps & {
+  title: string;
+  placeholder: string;
+  type: string;
+  name: string;
+};
